+++ src/main/java/blackjack/domain/player/Player.java
@@ -1,6 +1,6 @@
 package blackjack.domain.player;
 
-import blackjack.domain.Result;
+import blackjack.domain.result.Result;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import java.util.List;
