+++ src/main/java/blackjack/view/ResultView.java
@@ -1,7 +1,7 @@
 package blackjack.view;
 
-import blackjack.domain.GameResult;
-import blackjack.domain.Result;
+import blackjack.domain.result.GameResult;
+import blackjack.domain.result.Result;
 import blackjack.domain.player.Dealer;
 import blackjack.domain.player.Player;
 import blackjack.domain.player.User;
