+++ src/test/java/blackjack/domain/GameMachineTest.java
@@ -4,11 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.card.CardDeckGenerator;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardPattern;
 import blackjack.domain.player.Dealer;
-import blackjack.domain.player.User;
-import blackjack.domain.player.Users;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
