+++ src/test/java/blackjack/domain/card/CardDeckTest.java
@@ -1,18 +1,21 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.card.CardDeck;
+import blackjack.domain.card.CardDeckGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class CardDeckGeneratorTest {
+public class CardDeckTest {
 
     @Test
-    @DisplayName("52장의 카드를 포함하고 있는 카드덱을 정상 생성한다.")
-    void createCardDeck() {
+    @DisplayName("사용한 카드는 제거한다")
+    void drawCard() {
         final CardDeck cardDeck = CardDeckGenerator.createCardDeck();
-        final int expected = 52;
+        final int expected = 51;
 
+        cardDeck.selectCard();
         final int actual = cardDeck.size();
         assertThat(actual).isEqualTo(expected);
     }
