+++ src/test/java/blackjack/domain/money/BettingMoneyTest.java
@@ -1,8 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.money;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.money.BettingMoney;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
